Исправление проблем с ресурсами: перенос строк в модуль ui
- Перенесены все строковые ресурсы из модуля budget в модуль ui
- Удален файл strings.xml из модуля budget
- Обновлены импорты в StringProvider.kt и BudgetScreen.kt
- Добавлен AndroidManifest.xml для модуля budget
- Исправлены дублирующиеся строки (amount, error_linking_categories, success)
- Добавлены недостающие строки (error, ok, ok_button)
- Проект теперь соответствует правилам архитектуры

+++ domain/src/main/java/com/davidbugayov/financeanalyzer/domain/usecase/export/ExportTransactionsToCSVUseCase.kt
@@ -73,7 +73,7 @@ class ExportTransactionsToCSVUseCase(
                 }
             }
 
-            Timber.d(StringProvider.logCsvFileCreated(finalFile.absolutePath, finalFile.length()))
+            Timber.d(StringProvider.logCsvFileCreated(finalFile.absolutePath, finalFile.length().toInt()))
             return@withContext Result.Success(finalFile)
         } catch (e: IOException) {
             Timber.e(e, StringProvider.logExportError)
